repository

+++ app/src/main/java/com/viacheslav/movieguide/data/dto/MovieDetailsDto.kt
@@ -20,7 +20,7 @@ data class MovieDetailsDto(
     val backdropPath: String? = null,
 
 	@field:SerializedName("revenue")
-    val revenue: Int? = null,
+    val revenue: Long? = null,
 
 	@field:SerializedName("genres")
     val genres: List<GenreDto>,
